requirements.txt
azure-functions==1.16.0
openai==1.72.0
flask==2.3.3
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0

# LangChain and related packages
langchain==0.1.0
langchain-community==0.0.13
langsmith>=0.0.77,<0.1.0